+++ src/api/changesets/__tests__/_createOsmChangeXml.test.ts
@@ -32,7 +32,20 @@ describe("_createOsmChangeXml", () => {
           tags: { building: "yes" },
         },
       ],
-      delete: [],
+      delete: [
+        { type: "node", id: 15, ...boilerplate, lat: 0, lon: 0 },
+        { type: "relation", id: 101, ...boilerplate, members: [] },
+        { type: "node", id: 16, ...boilerplate, lat: 0, lon: 0 },
+        {
+          type: "way",
+          id: 3,
+          ...boilerplate,
+          nodes: [15, 16, 17, 18, 15],
+          tags: { building: "yes" },
+        },
+        { type: "node", id: 17, ...boilerplate, lat: 0, lon: 0 },
+        { type: "node", id: 18, ...boilerplate, lat: 0, lon: 0 },
+      ],
     };
     const xml = createOsmChangeXml(123, osmChange, {
       created_by: "me",
@@ -60,7 +73,21 @@ describe("_createOsmChangeXml", () => {
             <nd ref="10"/>
           </way>
         </modify>
-        <delete if-unused="true"/>
+        <delete if-unused="true">
+          <relation id="101" version="0" changeset="123"/>
+          <way id="3" version="0" changeset="123">
+            <tag k="building" v="yes"/>
+            <nd ref="15"/>
+            <nd ref="16"/>
+            <nd ref="17"/>
+            <nd ref="18"/>
+            <nd ref="15"/>
+          </way>
+          <node id="15" version="0" changeset="123" lat="0" lon="0"/>
+          <node id="16" version="0" changeset="123" lat="0" lon="0"/>
+          <node id="17" version="0" changeset="123" lat="0" lon="0"/>
+          <node id="18" version="0" changeset="123" lat="0" lon="0"/>
+        </delete>
       </osmChange>
       "
     `);